+++ main/weight.h
@@ -7,7 +7,7 @@
 #include <stdbool.h>
 
 void weight_init(settings_t *settings);
-float weight_get_latest(bool *available);
+double weight_get_latest(bool *available);
 uint32_t weight_get_latest_raw(bool *available);
 
 #endif // WEIGHT_H
